Added new tableview to display corporate blogs. Added storyboard items for seguing into article view from new table view.

+++ VMWareBlogs/VMArticleEntityUpdater.m
@@ -132,8 +132,8 @@ typedef enum {
                 int j = 0;
                 int order = 0;
                 int articleCount = 0;
-                int totalArticles = [sortedArticleArray count] == 0 ? 0 : ([sortedArticleArray count] -1);
-                                
+                int totalArticles = [sortedArticleArray count] == 0 ? 0 : (int)([sortedArticleArray count] -1);
+                
                 do {
                     Blog *blogEntry;
